requirements.txt
# AI Council Guide Creation Website Dependencies

# Quart Backend (v0.19)
quart==0.19.4
quart-cors==0.8.0
hypercorn==0.15.0

# MongoDB (v6) with PyMongo
pymongo==4.6.2
motor==3.3.2
dnspython==2.4.2

# LangChain (v0.2) for AI Orchestration
langchain==0.1.12
langchain-community==0.2.0
langchain-openai==0.0.8
langchain-anthropic==0.1.4
langchain-google-genai==0.0.11

# LLM Token Counting
tiktoken==0.6.0

# Async HTTP Requests
aiohttp==3.9.1

# PDF Generation
weasyprint==60.1

# Testing
pytest==8.0.0
pytest-flask==1.3.0

# Security
bcrypt==4.0.1
python-dotenv==1.0.1

# Development Tools
black==23.12.0
flake8==6.1.0

# Added from the code block
pydantic==2.6.4
typing-extensions==4.10.0 